% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{addSummaryChart}
\alias{addSummaryChart}
\title{Add PRIDE summary charts}
\usage{
addSummaryChart(res)
}
\arguments{
\item{res}{An object returned by msQCpipe function}
}
\description{
Add PRIDE summary charts in the technical replicate level
}
