% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msQC.R
\name{combineRun}
\alias{combineRun}
\title{Combine multiple results}
\usage{
combineRun(pepFiles, fasta, outPathFile, outdir, prefix)
}
\arguments{
\item{pepFiles}{peptideSummary files}

\item{fasta}{database file}

\item{outPathFile}{out file}

\item{outdir}{output directory}

\item{prefix}{output prefix}
}
\value{
A data.frame
}
\description{
Combine multiple results
}
\author{
Bo Wen \email{wenbo@genomics.cn}
}
