% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msQC.R
\name{createTargetDecoyDB}
\alias{createTargetDecoyDB}
\title{Create target-decoy database}
\usage{
createTargetDecoyDB(fa, outdb)
}
\arguments{
\item{fa}{target database}

\item{outdb}{output target-decoy database}
}
\value{
target-decoy database file name
}
\description{
Create target-decoy database
}
\author{
Bo Wen \email{wenbo@genomics.cn}
}
