% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msQC.R
\name{getEnzyme}
\alias{getEnzyme}
\title{Get the enzymes list}
\usage{
getEnzyme()
}
\value{
A data frame which contains all of the enzymes
}
\description{
Get the enzymes list
}
\author{
Bo Wen \email{wenbo@genomics.cn}
}
