% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msQC.R
\name{getMods}
\alias{getMods}
\title{Get the modification list}
\usage{
getMods()
}
\value{
A data frame which contains all of the modifications
}
\description{
Get the modification list
}
\author{
Bo Wen \email{wenbo@genomics.cn}
}
