% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{labelRatio}
\alias{labelRatio}
\title{Calculate the labeling efficiency of isobaric labeling data}
\usage{
labelRatio(ms = NULL, iClass = 1, delta = 0.05, plot = TRUE)
}
\arguments{
\item{ms}{MS/MS file.}

\item{iClass}{Isobaric tag class, 1=iTRAQ-8plex.}

\item{delta}{The mass error for reporter matching.}

\item{plot}{Logical value}
}
\value{
A vector object
}
\description{
Calculate the labeling efficiency of isobaric labeling data
}
\examples{
mgf.zip <- system.file("extdata/mgf.zip", package = "proteoQC")
unzip(mgf.zip)
a <- labelRatio("test.mgf")
}
\author{
Bo Wen \email{wenbo@genomics.cn}
}
