% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msQC.R
\name{loadmsQCres}
\alias{loadmsQCres}
\title{Load the result of \code{\link{msQCpipe}}}
\usage{
loadmsQCres(outdir)
}
\arguments{
\item{outdir}{The output directory of \code{\link{msQCpipe}}}
}
\description{
Load the result of \code{\link{msQCpipe}}
}
\examples{
zpqc <- system.file("extdata/qc.zip", package = "proteoQC")
unzip(zpqc)
qcres <- loadmsQCres("./qc")
}
\author{
Laurent Gatto \email{lg390@cam.ac.uk}, 
Bo Wen \email{wenbo@genomics.cn}
}
