% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msQC.R
\name{msQCpipe}
\alias{msQCpipe}
\title{The main function of msQC pipeline}
\usage{
msQCpipe(spectralist = NULL, fasta = "", outdir = "./", mode = "",
  miss = 2, enzyme = 1, varmod = NULL, fixmod = NULL, tol = 10,
  tolu = "ppm", itol = 0.6, itolu = "Daltons", threshold = 0.01,
  cpu = 0, xmx = 2, refine = TRUE, ntt = 1, ...)
}
\arguments{
\item{spectralist}{A file contains the experiment design or a single mgf file}

\item{fasta}{database file, must contain decoy sequences}

\item{outdir}{output directory}

\item{mode}{identification or quantification}

\item{miss}{max miss clevage}

\item{enzyme}{enzyme}

\item{varmod}{Variable modifications are those which may or may
not be present.}

\item{fixmod}{Fixed modifications are applied universally, to
every instance of the specified residue(s) or terminus.}

\item{tol}{The error window on experimental peptide mass values}

\item{tolu}{Units can be selected from: ppm, Daltons(also da or Da).}

\item{itol}{Error window for MS/MS fragment ion mass values.}

\item{itolu}{Units can be selected from: Daltons(also da or Da)}

\item{threshold}{FDR value for PSM}

\item{cpu}{Max number of cpu used}

\item{xmx}{JAVA -Xmx}

\item{refine}{Refine search for X!Tandem, default is TRUE.}

\item{ntt}{Semi-tryptic, 1; fully-tryptic, 2.}

\item{...}{Additional parameters passed to
\code{\link{read.table}} used to read the experimental design.}
}
\value{
A list which contains all of the information for data 
quality report generating
}
\description{
This function is designed to automate generating of 
target-decoy database, database searcing, post-processing and report 
generation.
}
\examples{
\dontrun{
library("rpx")
px <- PXDataset("PXD000864")
mgfs <- grep("mgf", pxfiles(px), value = TRUE)
mgfs <- grep("-0[5-6]-[1|2]", mgfs, value=TRUE)
mgffiles <- pxget(px, mgfs)
library("R.utils")
mgffiles <- sapply(mgffiles, gunzip)
## Generate the lightweight qc report, 
## trim the mgf files to 1/10 of their size.
trimMgf <- function(f, m = 1/10, overwrite = FALSE) {
  message("Reading ", f)
  x <- readLines(f)
  beg <- grep("BEGIN IONS", x)
  end <- grep("END IONS", x)
  n <- length(beg)
  message("Sub-setting to ", m)
  i <- sort(sample(n, floor(n * m)))
  k <- unlist(mapply(seq, from = beg[i], to = end[i]))
  if (overwrite) {
    unlink(f)
    message("Writing ", f)
    writeLines(x[k], con = f)
    return(f)
  } else {
    g <- sub(".mgf", "_small.mgf", f)
    message("Writing ", g)
    writeLines(x[k], con = g)
    return(g)
  }    
}
set.seed(1)
mgffiles <- sapply(mgffiles, trimMgf, overwrite = TRUE)
fas <- pxget(px, "TTE2010.zip")
fas <- unzip(fas)
design <- system.file("extdata/PXD000864-design.txt", package = "proteoQC")
read.table(design, header = TRUE)
qcres <- msQCpipe(spectralist = design,
                 fasta = fas, 
                 outdir = "./qc",
                 miss  = 0,
                 enzyme = 1, varmod = 2, fixmod = 1,
                 tol = 10, itol = 0.6, cpu = 2,
                 mode = "identification")
}
}
\author{
Bo Wen \email{wenbo@genomics.cn}
}
