% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plotBioRepVenn}
\alias{plotBioRepVenn}
\title{Venn plot in biological replicate level}
\usage{
plotBioRepVenn(res)
}
\arguments{
\item{res}{An object of msQCres}
}
\value{
The name of the figure
}
\description{
Venn plot in biological replicate level
}
