% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plotFractionIDResult}
\alias{plotFractionIDResult}
\title{Barplot in different level for each fraction}
\usage{
plotFractionIDResult(res, level = NA)
}
\arguments{
\item{res}{An object of msQCres}

\item{level}{1: total spectrum, 2: identified spectrum, 
3: identified peptide, 4: identified protein.}
}
\value{
The name of the figure
}
\description{
Barplot in different level for each fraction
}
