% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plotMS1Error}
\alias{plotMS1Error}
\title{plot MS1 mass error}
\usage{
plotMS1Error(res, plot.class = "ppm")
}
\arguments{
\item{res}{An object of msQCres}

\item{plot.class}{ppm or da}
}
\value{
The name of the figure
}
\description{
plot MS1 mass error
}
