% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plotMS2Error}
\alias{plotMS2Error}
\title{plot MS2 mass error}
\usage{
plotMS2Error(res)
}
\arguments{
\item{res}{An object of msQCres}
}
\value{
The name of the figure
}
\description{
plot MS2 mass error
}
