% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plotSampleIDResultErrorBar}
\alias{plotSampleIDResultErrorBar}
\title{Error barplot in different level for each fraction}
\usage{
plotSampleIDResultErrorBar(res, level = NA)
}
\arguments{
\item{res}{An object of parser result}

\item{level}{1: total spectrum, 2: identified spectrum, 
3: identified peptide, 4: identified protein.}
}
\value{
The name of the figure
}
\description{
Error Barplot in different level for each fraction
}
