% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plotSampleVenn}
\alias{plotSampleVenn}
\title{Venn plot in sample level}
\usage{
plotSampleVenn(res)
}
\arguments{
\item{res}{An object of msQCres}
}
\value{
The name of the figure
}
\description{
Venn plot in sample level
}
