% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plotTechRepVenn}
\alias{plotTechRepVenn}
\title{Venn plot in technical replicate level}
\usage{
plotTechRepVenn(res)
}
\arguments{
\item{res}{An object of msQCres}
}
\value{
The name of the figure
}
\description{
Venn plot in technical replicate level
}
