% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msQC.R
\name{print.msQCres}
\alias{print.msQCres}
\title{Print the information of msQCres object}
\usage{
\method{print}{msQCres}(x, ...)
}
\arguments{
\item{x}{A msQCres object}

\item{...}{Additional parameters}
}
\description{
Print the information of msQCres object
}
\examples{
zpqc <- system.file("extdata/qc.zip", package = "proteoQC")
unzip(zpqc)
qcres <- loadmsQCres("./qc")
print.msQCres(qcres)
}
\author{
Laurent Gatto \email{lg390@cam.ac.uk}, 
Bo Wen \email{wenbo@genomics.cn}
}
