% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{proteinGroup}
\alias{proteinGroup}
\title{Protein inference}
\usage{
proteinGroup(file = NULL, db = "", pepColName = "peptide",
  proColName = "protein", spectrumColName = "index", proSep = ";",
  outfile = NULL, xmx = 1)
}
\arguments{
\item{file}{A file containing the information of peptides to proteins.}

\item{db}{A protein database of fasta format.}

\item{pepColName}{The column name of peptide sequence.}

\item{proColName}{The column name of protein ID.}

\item{spectrumColName}{The column name of spectrum index.}

\item{proSep}{The separator of protein ID, default is "".}

\item{outfile}{The output file name of protein group result.}

\item{xmx}{JAVA -Xm}
}
\description{
Protein inference
}
\examples{
pep.zip <- system.file("extdata/pep.zip", package = "proteoQC")
unzip(pep.zip)
proteinGroup(file = "pep.txt", outfile = "pg.txt")
}
\author{
Bo Wen \email{wenbo@genomics.cn}
}
