% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{reportHTML}
\alias{reportHTML}
\title{HTML format report generator}
\usage{
reportHTML(res)
}
\arguments{
\item{res}{An object returned by \code{\link{msQCpipe}} function}
}
\value{
null
}
\description{
HTML format report generator
}
\examples{
zpqc <- system.file("extdata/qc.zip", package = "proteoQC")
unzip(zpqc)
qcres <- loadmsQCres("./qc")
html <- reportHTML(qcres)
}
\author{
Bo Wen \email{wenbo@genomics.cn}
}
