% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msQC.R
\name{runTandem}
\alias{runTandem}
\title{Run X!Tandem}
\usage{
runTandem(spectra = "", fasta = "", outdir = "./", outprefix = "",
  cpu = 1, enzyme = 1, xmx = 2, varmod = NULL, fixmod = NULL,
  refine = TRUE, ntt = 1, tol = 10, tolu = "ppm", itol = 0.6,
  itolu = "Daltons", miss = 1)
}
\arguments{
\item{spectra}{MS/MS peak list file}

\item{fasta}{database file}

\item{outdir}{output directory}

\item{outprefix}{output file prefix}

\item{cpu}{The number of CPU used for X!Tandem}

\item{enzyme}{The ID of enzyme used for database searching. 
See \code{\link{showEnzyme}}.}

\item{xmx}{Set for parameter of "Java -Xmx".}

\item{varmod}{Variable modifications used for database searching.
See \code{\link{showMods}}.}

\item{fixmod}{Fixed modifications used for database searching.
See \code{\link{showMods}}.}

\item{refine}{Refine search, default is TRUE}

\item{ntt}{Default is 1}

\item{tol}{The error window on experimental peptide mass values}

\item{tolu}{Units can be selected from: ppm, Daltons.}

\item{itol}{Error window for MS/MS fragment ion mass values.}

\item{itolu}{Units can be selected from: Daltons}

\item{miss}{Max miss clevage}
}
\value{
a file path
}
\description{
Run X!Tandem

run X!Tandem
}
\author{
Bo Wen \email{wenbo@genomics.cn}
}
