% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msQC.R
\name{showEnzyme}
\alias{showEnzyme}
\title{Shown all enzymes}
\usage{
showEnzyme()
}
\value{
A data frame which contains all of the enzymes
}
\description{
Shown all enzymes
}
\examples{
showEnzyme()
}
\author{
Bo Wen \email{wenbo@genomics.cn}
}
