% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msQC.R
\name{showMods}
\alias{showMods}
\title{Shown all modifications}
\usage{
showMods()
}
\value{
A data frame which contains all of the modifications
}
\description{
Shown all modifications
}
\examples{
showMods()
}
\author{
Bo Wen \email{wenbo@genomics.cn}
}
