% Generated by roxygen2 (4.0.1): do not edit by hand
\name{calcMSQCMetrics}
\alias{calcMSQCMetrics}
\title{Calculate the MS1 and MS2 level QC metrics}
\usage{
calcMSQCMetrics(spectraList = NULL, cpu = 2, outdir = "./")
}
\arguments{
\item{spectraList}{An experiment design input file}

\item{outdir}{Output directory}

\item{cpu}{Number of cpu used}
}
\description{
Calculate the MS1 level QC metrics
}
\author{
Bo Wen \email{wenbo@genomics.cn}
}

