% Generated by roxygen2 (4.0.2): do not edit by hand
\name{chargeStat}
\alias{chargeStat}
\title{Charge distribution}
\usage{
chargeStat(mgf = NULL)
}
\arguments{
\item{mgf}{A file of mgf.}
}
\value{
A vector object
}
\description{
Read the charge information from mgf file
}
\examples{
mgf.zip <- system.file("extdata/mgf.zip", package = "proteoQC")
unzip(mgf.zip)
charge <- chargeStat("test.mgf")
}
\author{
Bo Wen \email{wenbo@genomics.cn}
}

