% Generated by roxygen2 (4.0.1): do not edit by hand
\name{cntStat}
\alias{cntStat}
\title{contaminants stat}
\usage{
cntStat(res)
}
\arguments{
\item{res}{An object of msQCres}
}
\value{
A data.frame will be shown in HTML report
}
\description{
Common Contaminants in Proteomics Mass Spectrometry Experiments
}

