% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{labelRatio}
\alias{labelRatio}
\title{Calculate the labeling efficiency of isobaric labeling data}
\usage{
labelRatio(ms = NULL, reporter = 1, plot = TRUE)
}
\arguments{
\item{ms}{MS/MS file.}

\item{reporter}{Isobaric tag class, 1=iTRAQ-4plex, 2=iTRAQ-8plex, 3=TMT-6plex.
4=TMT-10plex.}

\item{plot}{Logical value}
}
\value{
A vector object
}
\description{
Calculate the labeling efficiency of isobaric labeling data
}
\examples{
mgf.zip <- system.file("extdata/mgf.zip", package = "proteoQC")
unzip(mgf.zip)
a <- labelRatio("test.mgf",reporter=2)
}
\author{
Bo Wen \email{wenbo@genomics.cn}
}
