% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plotBioRepVenn}
\alias{plotBioRepVenn}
\title{Venn plot in biological replicate level}
\usage{
plotBioRepVenn(res)
}
\arguments{
\item{res}{An object of msQCres}
}
\value{
The name of the figure
}
\description{
Venn plot in biological replicate level
}

