% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plotMS2Error_obsolete}
\alias{plotMS2Error_obsolete}
\title{plot MS2 mass error}
\usage{
plotMS2Error_obsolete(res)
}
\arguments{
\item{res}{An object of msQCres}
}
\value{
The name of the figure
}
\description{
plot MS2 mass error
}

