% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plotSampleVenn}
\alias{plotSampleVenn}
\title{Venn plot in sample level}
\usage{
plotSampleVenn(res)
}
\arguments{
\item{res}{An object of msQCres}
}
\value{
The name of the figure
}
\description{
Venn plot in sample level
}

