% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plotTechRepVenn}
\alias{plotTechRepVenn}
\title{Venn plot in technical replicate level}
\usage{
plotTechRepVenn(res)
}
\arguments{
\item{res}{An object of msQCres}
}
\value{
The name of the figure
}
\description{
Venn plot in technical replicate level
}

