% Generated by roxygen2 (4.0.1): do not edit by hand
\name{showEnzyme}
\alias{showEnzyme}
\title{Shown all enzymes}
\usage{
showEnzyme()
}
\value{
A data frame which contains all of the enzymes
}
\description{
Shown all enzymes
}
\examples{
showEnzyme()
}
\author{
Bo Wen \email{wenbo@genomics.cn}
}

